refactor(validator): streamline validation process and improve error handling refactor(files): modularize utility functions for better organization fix(import-gtfs): correct error handling in GTFS import process
