+++ packages/files/src/utils/is-browser.ts
@@ -1,0 +1,4 @@
+/* * */
+
+declare const window: typeof globalThis;
+export const isBrowser = typeof window?.document !== 'undefined';
