+++ packages/files/src/utils/normalize-file.content.ts
@@ -1,0 +1,19 @@
+/* * */
+
+import { isBrowser } from '@/utils/is-browser.js';
+
+/**
+ * Converts the input file data into an ArrayBuffer or Uint8Array suitable for JSZip.
+ * @param content A File (browser) or Buffer/Uint8Array (Node.js)
+ */
+export async function normalizeFileContent(content: Buffer | File | Uint8Array): Promise<ArrayBuffer | Uint8Array> {
+	if (isBrowser && content instanceof File) {
+		return await content.arrayBuffer();
+	}
+
+	if (content instanceof Buffer || content instanceof Uint8Array) {
+		return content;
+	}
+
+	throw new TypeError('Unsupported file content type for zipping.');
+}
